# openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas Backend API
  description: Personal finance ETL + API
  version: 0.3.0
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Login
      description: |
        Authenticate with username and password (form data).
        Returns a Bearer JWT token valid for 8 hours.
      operationId: login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Register with invite token
      description: |
        Create a new user account using a valid one-time invite token generated
        by an admin. The token is invalidated after use.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created and logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid, used, or expired invite token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

  # ──────────────────────────────────────────────
  # ADMIN
  # ──────────────────────────────────────────────
  /admin/invite:
    post:
      tags: [admin]
      summary: Create an invite token
      description: |
        Generates a one-time invite token valid for 7 days.
        Share it with the person you want to invite — they use it at `POST /auth/register`.
      operationId: createInvite
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Invite token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      operationId: listUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /admin/users/{user_id}/promote:
    patch:
      tags: [admin]
      summary: Promote user to admin
      operationId: promoteUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User promoted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

  # ──────────────────────────────────────────────
  # ACCOUNTS
  # ──────────────────────────────────────────────
  /accounts/{account_id}:
    get:
      tags: [accounts]
      summary: Get account by ID
      operationId: getAccount
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

    put:
      tags: [accounts]
      summary: Update an account
      operationId: updateAccount
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

    delete:
      tags: [accounts]
      summary: Delete an account
      description: |
        Deletes the account and cascade-deletes all associated source files,
        transactions, and raw rows.
      operationId: deleteAccount
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /accounts:
    get:
      tags: [accounts]
      summary: List accounts for the current user
      description: |
        Returns only the accounts owned by the authenticated user.
        Admin users see all accounts.
      operationId: listAccounts
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OwnerFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

    post:
      tags: [accounts]
      summary: Create a new account
      description: |
        Creates an account owned by the authenticated user.
        The `owner` field is a free-form label (e.g. "PACHO", "LU", or any string).
      operationId: createAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
            example:
              bank_name: BANCOLOMBIA
              account_name: Cuenta de Ahorros
              owner: PACHO
              account_type: AHORROS
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'

  # ──────────────────────────────────────────────
  # FILES
  # ──────────────────────────────────────────────
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AccountId:
      name: account_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Account UUID
    OwnerFilter:
      name: owner
      in: query
      required: false
      schema:
        type: string
      description: Filter by owner label (e.g. "PACHO", "LU")
...
  schemas:
...
    BankEnum:
      type: string
      enum: [BANCOLOMBIA, RAPPI, FALABELLA, NEQUI]
    AccountTypeEnum:
      type: string
      enum: [CREDITO, DEBITO, AHORROS]
...
    RegisterRequest:
      type: object
      required: [username, password, invite_token]
      properties:
        username:
          type: string
          example: lupa
        password:
          type: string
          format: password
        invite_token:
          type: string
          format: uuid
          description: One-time invite token generated by an admin
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        is_admin:
          type: boolean
          description: Whether the authenticated user has admin privileges

    # ── Admin ──────────────────────────────────
    InviteResponse:
      type: object
      properties:
        invite_token:
          type: string
          format: uuid
          description: One-time token to share with the invitee
        expires_at:
          type: string
          format: date-time
          description: Token expiry (7 days from creation)
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Accounts ───────────────────────────────
    AccountUpdate:
      type: object
      properties:
        bank_name:
          $ref: '#/components/schemas/BankEnum'
        account_name:
          type: string
        owner:
          type: string
          description: Free-form owner label (e.g. "PACHO", "LU")
        account_type:
          $ref: '#/components/schemas/AccountTypeEnum'
    AccountCreate:
      type: object
      required: [bank_name, account_name, owner, account_type]
      properties:
        bank_name:
          $ref: '#/components/schemas/BankEnum'
        account_name:
          type: string
          example: Cuenta de Ahorros
        owner:
          type: string
          description: Free-form owner label (e.g. "PACHO", "LU")
          example: PACHO
        account_type:
          $ref: '#/components/schemas/AccountTypeEnum'
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bank_name:
          $ref: '#/components/schemas/BankEnum'
        account_name:
          type: string
        owner:
          type: string
          description: Free-form owner label
        account_type:
          $ref: '#/components/schemas/AccountTypeEnum'
        user_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of the user who owns this account

    # ── Files ──────────────────────────────────
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string